modified:   src/main/java/net/ronondex2009/essence_of_harmony/event/Events.java 	modified:   src/main/java/net/ronondex2009/essence_of_harmony/spell/ModSpells.java 	modified:   src/main/java/net/ronondex2009/essence_of_harmony/spell/custom/BreakBlockSpell.java 	modified:   src/main/java/net/ronondex2009/essence_of_harmony/spell/custom/DirectionOperator.java 	modified:   src/main/java/net/ronondex2009/essence_of_harmony/spell/custom/LocationOperator.java 	modified:   src/main/java/net/ronondex2009/essence_of_harmony/spell/custom/RaycastBlockOperator.java 	new file:   src/main/java/net/ronondex2009/essence_of_harmony/spell/custom/ShowEssenceOperator.java 	modified:   src/main/java/net/ronondex2009/essence_of_harmony/util/Spell.java 	modified:   src/main/java/net/ronondex2009/essence_of_harmony/world/essenceMap.java

+++ src/main/java/net/ronondex2009/essence_of_harmony/spell/custom/BreakBlockSpell.java
@@ -2,9 +2,13 @@ package net.ronondex2009.essence_of_harmony.spell.custom;
 
 import java.util.List;
 
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Blocks;
+import net.ronondex2009.essence_of_harmony.block.ModBlocks;
 import net.ronondex2009.essence_of_harmony.item.ModItems;
 import net.ronondex2009.essence_of_harmony.sound.ModSounds;
 import net.ronondex2009.essence_of_harmony.spell.symbols.VectorSymbol;
@@ -12,8 +16,8 @@ import net.ronondex2009.essence_of_harmony.util.AbstractSymbol;
 import net.ronondex2009.essence_of_harmony.util.Spell;
 import net.ronondex2009.essence_of_harmony.util.notes;
 
-public class BreakBlockSpell extends Spell {
-    
+public class BreakBlockSpell extends Spell 
+{
     @Override
     public boolean runSpell(List<AbstractSymbol> stack, Player player, Level level) 
     {        
@@ -21,8 +25,14 @@ public class BreakBlockSpell extends Spell {
         if(!(stack.get(stack.size()-1).getSymbolType().equals("Vector"))) return false;
         
         VectorSymbol location = (VectorSymbol) stack.get(stack.size()-1);
-        level.destroyBlock(new BlockPos(location.getX(), location.getY(), location.getZ()), true);
         stack.remove(stack.size()-1);
+        if(level.getBlockState(new BlockPos(location.getX(), location.getY(), location.getZ())).getBlock().defaultDestroyTime() > power/1.5 && (power==3 && level.getBlockState(new BlockPos(location.getX(), location.getY(), location.getZ())).getBlock()==ModBlocks.NOTIUM_ORE.get()))
+        {
+            if(level.isClientSide)
+                player.sendSystemMessage(Component.literal("The spell wasn't powerful enough to destroy the block!").withStyle(ChatFormatting.RED));
+            return false;
+        }
+        level.destroyBlock(new BlockPos(location.getX(), location.getY(), location.getZ()), true);
 
         player.playSound(ModSounds.CAST_SPELL.get());
         return true;
@@ -42,5 +52,7 @@ public class BreakBlockSpell extends Spell {
         allowedInstruments.add(ModItems.OCARINA.get());
         allowedInstruments.add(ModItems.GUITAR.get());
         allowedInstruments.add(ModItems.OVERDRIVE_GUITAR.get());
+        baseEssenceUsage = 0.1f;
     }
+
 }
